LayoutTests/http/tests/security/cross-origin-window-property-access-expected.txt

Tests that using another window's property getter does not bypass cross-origin checks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getOwnPropertyDescriptor(window, "document").get.call(crossOriginWindow) threw exception TypeError: undefined is not an object (evaluating 'Object.getOwnPropertyDescriptor(window, "document").get.call').
PASS Object.getOwnPropertyDescriptor(window, "location").get.call(crossOriginWindow) threw exception TypeError: undefined is not an object (evaluating 'Object.getOwnPropertyDescriptor(window, "location").get.call').
PASS successfullyParsed is true

TEST COMPLETE
